# [Enhancement] support local runtime filter in cross join