--- a/be/src/exec/pipeline/crossjoin/cross_join_context.h
+++ b/be/src/exec/pipeline/crossjoin/cross_join_context.h
@@ -8,13 +8,32 @@
 #include "column/chunk.h"
 #include "column/vectorized_fwd.h"
 #include "exec/pipeline/context_with_dependency.h"
+#include "exprs/expr_context.h"
+#include "runtime/runtime_state.h"
+
+namespace starrocks::vectorized {
+class RuntimeFilterBuildDescriptor;
+}
 
 namespace starrocks::pipeline {
+class RuntimeFilterHub;
+struct CrossJoinContextParams {
+    int32_t num_right_sinkers;
+    int32_t plan_node_id;
+    std::vector<ExprContext*> filters;
+    RuntimeFilterHub* rf_hub;
+    std::vector<vectorized::RuntimeFilterBuildDescriptor*> rf_descs;
+};
 
 class CrossJoinContext final : public ContextWithDependency {
 public:
-    explicit CrossJoinContext(const int32_t num_right_sinkers)
-            : _num_right_sinkers(num_right_sinkers), _build_chunks(num_right_sinkers) {}
+    explicit CrossJoinContext(CrossJoinContextParams params)
+            : _num_right_sinkers(params.num_right_sinkers),
+              _plan_node_id(params.plan_node_id),
+              _build_chunks(_num_right_sinkers),
+              _conjuncts_ctx(std::move(params.filters)),
+              _rf_hub(params.rf_hub),
+              _rf_descs(std::move(params.rf_descs)) {}
 
     void close(RuntimeState* state) override {}
 
@@ -31,14 +50,22 @@ public:
         _build_chunks[sinker_id] = build_chunk;
     }
 
-    void finish_one_right_sinker() { _num_finished_right_sinkers.fetch_add(1, std::memory_order_release); }
-
-    bool is_right_finished() const {
-        return _num_finished_right_sinkers.load(std::memory_order_acquire) == _num_right_sinkers;
+    Status finish_one_right_sinker(RuntimeState* state) {
+        if (_num_right_sinkers - 1 == _num_finished_right_sinkers.fetch_add(1)) {
+            RETURN_IF_ERROR(_init_runtime_filter(state));
+            _all_right_finished.store(true, std::memory_order_release);
+        }
+        return Status::OK();
     }
 
+    bool is_right_finished() const { return _all_right_finished.load(std::memory_order_acquire); }
+
 private:
+    Status _init_runtime_filter(RuntimeState* state);
+
     const int32_t _num_right_sinkers;
+
+    const int32_t _plan_node_id;
     // A CrossJoinLeftOperator waits for all the CrossJoinRightSinkOperators to be finished,
     // and then reads _build_chunks written by the CrossJoinRightSinkOperators.
     // _num_finished_right_sinkers is used to ensure CrossJoinLeftOperator can see all the parts
@@ -47,6 +74,16 @@ private:
 
     // _build_chunks[i] contains all the rows from i-th CrossJoinRightSinkOperator.
     std::vector<vectorized::ChunkPtr> _build_chunks;
+
+    // finished flags
+    std::atomic_bool _all_right_finished = false;
+
+    // conjuncts in cross join, used for generate runtime_filter
+    std::vector<ExprContext*> _conjuncts_ctx;
+
+    RuntimeFilterHub* _rf_hub;
+
+    std::vector<vectorized::RuntimeFilterBuildDescriptor*> _rf_descs;
 };
 
 } // namespace starrocks::pipeline