--- a/be/src/exec/vectorized/project_node.cpp
+++ b/be/src/exec/vectorized/project_node.cpp
@@ -9,6 +9,7 @@
 
 #include "column/binary_column.h"
 #include "column/chunk.h"
+#include "column/column.h"
 #include "column/column_helper.h"
 #include "column/nullable_column.h"
 #include "column/vectorized_fwd.h"
@@ -217,15 +218,14 @@ void ProjectNode::push_down_predicate(RuntimeState* state, std::list<ExprContext
             continue;
         }
 
-        DCHECK(nullptr != dynamic_cast<vectorized::ColumnRef*>(ctx->root()->get_child(0)));
-        auto column = ((vectorized::ColumnRef*)ctx->root()->get_child(0));
+        auto column = down_cast<vectorized::ColumnRef*>(ctx->root()->get_child(0));
 
         for (int i = 0; i < _slot_ids.size(); ++i) {
             if (_slot_ids[i] == column->slot_id() && _expr_ctxs[i]->root()->is_slotref()) {
-                DCHECK(nullptr != dynamic_cast<vectorized::ColumnRef*>(_expr_ctxs[i]->root()));
-                auto ref = ((vectorized::ColumnRef*)_expr_ctxs[i]->root());
+                auto ref = down_cast<vectorized::ColumnRef*>(_expr_ctxs[i]->root());
                 column->set_slot_id(ref->slot_id());
                 column->set_tuple_id(ref->tuple_id());
+                break;
             }
         }
     }