@stdpain 2022-05-18
> Consider existing solutions. And how can you support the right table to return multiple rows of data？should you push down min/max?

we can support it later.