@wanpengfei-git 2022-05-18
#### [FE PR Coverage check]

:heart_eyes: **pass** : 58 / 64 (90.62%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/planner/PlanFragment.java|7|9|77.78%|
|:large_blue_circle:|com/starrocks/planner/CrossJoinNode.java|40|44|90.91%|
|:large_blue_circle:|com/starrocks/sql/plan/PlanFragmentBuilder.java|7|7|100.00%|
|:large_blue_circle:|com/starrocks/planner/JoinNode.java|3|3|100.00%|
|:large_blue_circle:|com/starrocks/planner/PlanNode.java|1|1|100.00%|