@stdpain 2022-05-20
run starrocks_fe_unittest