@wanpengfei-git 2022-05-20
#### [FE PR Coverage check]

:heart_eyes: **pass** : 62 / 67 (92.54%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|com/starrocks/planner/PlanFragment.java|7|9|77.78%|
|:large_blue_circle:|com/starrocks/planner/RuntimeFilterDescription.java|4|5|80.00%|
|:large_blue_circle:|com/starrocks/planner/CrossJoinNode.java|41|43|95.35%|
|:large_blue_circle:|com/starrocks/sql/plan/PlanFragmentBuilder.java|7|7|100.00%|
|:large_blue_circle:|com/starrocks/planner/JoinNode.java|2|2|100.00%|
|:large_blue_circle:|com/starrocks/planner/PlanNode.java|1|1|100.00%|